backend/requirements.txt
fastapi==0.103.2
meilisearch==0.28.4
uvicorn==0.23.2
openai==0.28.1
python-dotenv==1.0.0

